requirements.txt
crewai>=1.9.0
pydantic>=2.0.0
PyYAML>=6.0.0
tavily-python>=0.6.0
httpx>=0.27.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
sse-starlette>=2.1.0
